Separate shed_util_common into Galaxy and TS variants.
